+++ src/exception/index.ts
@@ -1,0 +1,2 @@
+export { NotFoundException } from './not-found-exception';
+export { UserErrorsException } from './user-errors-exception';