+++ src/index.ts
@@ -3,7 +3,7 @@ export { MetaobjectRepository } from './metaobject-repository';
 export { metafieldRepository, MetafieldRepository } from './metafield-repository';
 export { storefrontTokenRepository, StorefrontTokenRepository } from './storefront-token-repository';
 
-export { UserErrorsException } from './exception/user-errors-exception';
+export { NotFoundException, UserErrorsException } from './exception';
 
 export type { FieldDefinition, DefinitionSchemaEntry, DefinitionSchema } from './types/definitions';
 export type { InferObjectType, OnPopulateFunc } from './types/metaobject-repository';